+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ### Features
 
  * Cut out PlantUML's formatter to interface. This is preparation for Mermaid support.
+ * add `mermaid` option.
+ * add Mermaid formatter.
 
 ## v1.1.0 (2023-04-24)